changing name of images directory
