+++ csss-site/src/events/static/events_static/frosh/2015/images/house/BASIC_tronLogo.png
