+++ csss-site/src/events/static/events_static/frosh/2015/images/logos/Simba Logo white on transparent triangle with transparent S.png
