+++ csss-site/src/events/static/events_static/frosh/2015/images/logos/archiact_type_logo_WHT_large.png
