+++ csss-site/src/events/static/events_static/frosh/2015/images/misc/blueDisk_landscape.jpg
