+++ csss-site/src/events/static/events_static/frosh/2015/images/team/kenneth.jpg
