+++ csss-site/src/events/templates/events/frosh/2015/index.html
@@ -8,23 +8,23 @@
     <meta name="author" content="Bosco & Corbett">
     <meta name="creator" content="Bosco Noronha, nbosco.me">
     <meta name="attribution_parties" content="SHIELD free bootstrap theme - blacktie.co licensed under CC3">
-    <link rel="shortcut icon" href="{%  static 'events_static/frosh/2015_v2/img/team/favicon.ico' %}">
+    <link rel="shortcut icon" href="{%  static 'events_static/frosh/2015/images/favicon.ico' %}">
 
     <title>Computing Science Student Society Frosh 2015, Sponsored by Simba Technologies</title>
 
     <!-- Bootstrap core CSS -->
-    <link href="{%  static 'events_static/frosh/2015_v2/css/bootstrap.css' %}" rel="stylesheet">
+    <link href="{%  static 'events_static/frosh/2015/css/bootstrap.css' %}" rel="stylesheet">
 
     <!-- Custom styles for this template -->
-    <link href="{%  static 'events_static/frosh/2015_v2/css/soon.css' %}" rel="stylesheet">
+    <link href="{%  static 'events_static/frosh/2015/css/soon.css' %}" rel="stylesheet">
 
     <!-- Overrides for everything -->
-    <link href="{%  static 'events_static/frosh/2015_v2/css/overrides.css' %}" rel="stylesheet">
+    <link href="{%  static 'events_static/frosh/2015/css/overrides.css' %}" rel="stylesheet">
 
     <!-- HTML5 shim and Respond.js IE8 support of HTML5 elements and media queries -->
     <!--[if lt IE 9]>
-      <script src="{%  static 'events_static/frosh/2015_v2/js/html5shiv.js' %}"></script>
-      <script src="{%  static 'events_static/frosh/2015_v2/js_v2/respond.min.js' %}"></script>
+      <script src="{%  static 'events_static/frosh/2015/js/html5shiv.js' %}"></script>
+      <script src="{%  static 'events_static/frosh/2015/js_v2/respond.min.js' %}"></script>
     <![endif]-->
 
     <link href='https://fonts.googleapis.com/css?family=Lato:300,400,700,300italic,400italic' rel='stylesheet' type='text/css'>
@@ -71,7 +71,7 @@
         <!-- START TIMER -->
 
         <div class="row">
-            <img class="img-responsive center-block" src="{%  static 'events_static/frosh/2015_v2/img/misc/blueDisk.jpg' %}">
+            <img class="img-responsive center-block" src="{%  static 'events_static/frosh/2015/images/misc/blueDisk.jpg' %}">
         </div>
 
         <div class="row">
@@ -86,7 +86,7 @@
         <div class="row">
             <div class="col-lg-4 col-lg-offset-4 mt centered">
                 <button type="button" class="btn btn-info" onclick='window.location.href = "https://docs.google.com/forms/d/1KcNNdA1Iy11nPLnjwRf1H5aIw_VuB8cx5CHDvjtShYI/viewform?usp=send_form";'>Sign Up Now!</button>
-                <button type="button" class="btn btn-default" onclick='window.location.href = "{{ URL_ROOT }}events/frosh/2015_v2/frosh";'><a href="{{ URL_ROOT }}events/frosh/2015_v2/frosh">Enter Website!</a></button>
+                <button type="button" class="btn btn-default" onclick='window.location.href = "{{ URL_ROOT }}events/frosh/2015/frosh";'><a href="{{ URL_ROOT }}events/frosh/2015/frosh">Enter Website!</a></button>
             </div>
         </div>
         <!-- END TIMER -->
@@ -94,7 +94,7 @@
         <div class="row">
             <h1>Sponsored by: </h1>
             <div>
-                <img class="img-responsive center-block" src="{%  static 'events_static/frosh/2015_v2/img/logos/Simba Logo white on transparent triangle with transparent S.png' %}">
+                <img class="img-responsive center-block" src="{%  static 'events_static/frosh/2015/images/logos/Simba Logo white on transparent triangle with transparent S.png' %}">
             </div>
         </div>
     </div><!-- End Container -->
@@ -109,16 +109,16 @@
         window.location.href = "https://docs.google.com/forms/d/1KcNNdA1Iy11nPLnjwRf1H5aIw_VuB8cx5CHDvjtShYI/viewform?usp=send_form";
     }
     function mainsiteRedirect() {
-        window.location.href = "{{ URL_ROOT }}events/frosh/2015_v2/frosh";
+        window.location.href = "{{ URL_ROOT }}events/frosh/2015/frosh";
     }
 
 </script>
-<script src="{%  static 'events_static/frosh/2015_v2/js/jquery.min.js' %}"></script>
-<script type="text/javascript" src="{%  static 'events_static/frosh/2015_v2/js/modernizr.custom.js' %}"></script>
-<script src="{%  static 'events_static/frosh/2015_v2/js/bootstrap.js' %}"></script>
-<script src="{%  static 'events_static/frosh/2015_v2/js/soon/plugins.js' %}"></script>
-<script src="{%  static 'events_static/frosh/2015_v2/js/soon/jquery.themepunch.revolution.min.js' %}"></script>
-<script src="{%  static 'events_static/frosh/2015_v2/js/soon/custom.js' %}"></script>
+<script src="{%  static 'events_static/frosh/2015/js/jquery.min.js' %}"></script>
+<script type="text/javascript" src="{%  static 'events_static/frosh/2015/js/modernizr.custom.js' %}"></script>
+<script src="{%  static 'events_static/frosh/2015/js/bootstrap.js' %}"></script>
+<script src="{%  static 'events_static/frosh/2015/js/soon/plugins.js' %}"></script>
+<script src="{%  static 'events_static/frosh/2015/js/soon/jquery.themepunch.revolution.min.js' %}"></script>
+<script src="{%  static 'events_static/frosh/2015/js/soon/custom.js' %}"></script>
 </body>
 <!-- END BODY -->
 </html>
